+++ Soggins/ExternalAPIs/IAppleArtifactApi.cs
@@ -1,0 +1,7 @@
+﻿namespace Shockah.Soggins;
+
+public interface IAppleArtifactApi
+{
+	public delegate CardAction SingleActionProvider(State state);
+	public void SetPaletteAction(Deck deck, SingleActionProvider action, Tooltip description);
+}