+++ Soggins/I18n.cs
@@ -106,6 +106,8 @@ internal static class I18n
 	public static string ExeCardTextA => $"Get 1 <c=status>SMUG</c>. Add 1 of 2 <c=cardtrait>discnt, temp</c> <c={SogginsColor}>{SogginsName}</c> cards to your hand.";
 	public static string ExeCardTextB => $"<c=status>SMUG</c> = 0. Add 1 of 3 <c=cardtrait>discnt, temp</c> <c={SogginsColor}>{SogginsName}</c> cards to your hand.";
 
+	public static string PaletteTooltip => "<c=B79CE5>SOGGINS PIGMENT</c>\nYour next colorless card also adds a random <c=card>Halfhearted Apology</c> to your hand.";
+
 	public static readonly string[] ApologyFlavorTexts = [
 		"I'm sorry you feel that way",
 		"Sowwee",